backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
openai==1.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.4.2
manimgl==1.6.1
pymongo==4.5.0
requests==2.31.0
Werkzeug==2.3.7
boto3==1.34.61